requirements.txt
aiosqlite==0.18.0
anyio==3.6.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crypto==1.4.1
cryptography==40.0.1
databases==0.7.0
ecdsa==0.18.0
fastapi==0.80.0
greenlet==2.0.2
h11==0.13.0
httptools==0.4.0
idna==3.3
Naked==0.1.32
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.2
PyJWT==2.6.0
python-dotenv==0.20.0
python-jose==3.3.0
pytube==12.1.2
PyYAML==6.0
requests==2.28.2
rsa==4.9
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.47
starlette==0.19.1
typing-extensions==4.3.0
urllib3==1.26.15
uvicorn==0.18.3
watchfiles==0.16.1
websockets==10.3
